package/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "selfclaw"
version = "1.0.0"
description = "Selfclaw - AI Agent System based on GLM-5"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "raychiu0202", email = "your-email@example.com"}
]
maintainers = [
    {name = "raychiu0202"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agent", "chatbot", "llm", "glm", "terminal", "automation"]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "zhipuai>=2.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
selfclaw = "selfclaw.cli:main"

[project.urls]
"Bug Reports" = "https://github.com/raychiu0202/selfclaw/issues"
"Source" = "https://github.com/raychiu0202/selfclaw"
"Documentation" = "https://github.com/raychiu0202/selfclaw/docs"

[tool.setuptools]
packages = ["selfclaw"]

[tool.setuptools.package-data]
selfclaw = ["*.yaml", "*.yml", "*.json", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]